included ULC, updated plotting
Included unconstrained linear combination as a weighting scheme, even though it still needs debugging, as it provides probabilities greater than 1. This can be seen when "desired_forecast"="M1+", "desired_metric"="breir", and "desired_weighting"="unconstrained". This can be tested in ens_test.py.

Also updated the plotting functionality so that the diagrams may be plotted on desired axis/set of axes. This is used in my thesis. As well as this, have added different marker  formats depending on weighting scheme of ensemble.